requirements.txt
streamlit==1.33.0
google-generativeai==0.3.2
transformers==4.40.0
torch==2.3.0+cpu
sentencepiece==0.2.0
protobuf==4.25.3
tqdm==4.66.2